# swagger.json
{"openapi":"3.0.1","info":{"title":"Baeker-Study","description":"Baeker 스터디의 API 명세서 입니다.","license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.0"},"servers":[{"url":"http://localhost:8082"}],"tags":[{"name":"My Study","description":"My Study - CRUD"},{"name":"Study","description":"Study - CRUD"}],"paths":{"/api/studyrule/v1/{studyruleid}":{"put":{"tags":["StudyRule-수정"],"summary":"스터디 규칙 전부 변경","description":"body 에 이름(name), 소개(about), 규칙(ruleId) 입력","operationId":"modifyStudyRule","parameters":[{"name":"studyruleid","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyStudyRuleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataStudyRuleDto"}}}}}},"patch":{"tags":["StudyRule-수정"],"summary":"스터디규칙 개별 변경","description":"이름(name), 소개(about), 규칙Id(ruleId) 원하는 값을 Key:Value 형태로 입력","operationId":"modifyStudyRule_1","parameters":[{"name":"studyruleid","in":"path","description":"수정하고싶은 StudyRuleId 입력","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"},"description":"key:value 형태로(ex: name:이름) 전달"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataStudyRuleDto"}}}}}}},"/api/studyrule/v1/studyrules":{"post":{"tags":["StudyRule-생성"],"summary":"스터디규칙 생성","description":"StudyRule 생성","operationId":"createStudyRule","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStudyRuleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataCreateStudyRuleResponse"}}}}}}},"/api/study/v1/xp":{"post":{"tags":["Study"],"summary":"XP 추가","operationId":"addXp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddXpReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/study/v1/update":{"post":{"tags":["Study"],"summary":"업데이트 name / about / capacity","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/study/v1/leader":{"post":{"tags":["Study"],"summary":"업데이트 leader","operationId":"updateLeader","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLeaderReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/study/v1/create":{"post":{"tags":["Study"],"summary":"Study 생성","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/my-study/v1/msg":{"post":{"tags":["My Study"],"summary":"가입, 초대 메시지 업데이트","operationId":"modifyMsg","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyMsgDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/my-study/v1/join":{"post":{"tags":["My Study"],"summary":"Study 에 가입 신청","operationId":"join","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinMyStudyReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/my-study/v1/invite":{"post":{"tags":["My Study"],"summary":"Study 로 초대","operationId":"invite","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteMyStudyReqDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/my-study/v1/accept":{"post":{"tags":["My Study"],"summary":"가입, 초대 승인","operationId":"accept","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcceptDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/studyrule/v1/search":{"get":{"tags":["StudyRule-조회"],"summary":"스터디 규칙 전체조회","description":"모든 StudyRule 을 조회 합니다.","operationId":"searchStudyRule","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataListStudyRuleDto"}}}}}}},"/api/studyrule/v1/search/{studyruleid}":{"get":{"tags":["StudyRule-조회"],"summary":"스터디 규칙 개별조회","description":"한개씩 조회합니다.","operationId":"searchStudyRuleId","parameters":[{"name":"studyruleid","in":"path","description":"조회 하고싶은 StudyRuleId 입력","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsDataStudyRuleDto"}}}}}}},"/api/study/v1/name":{"get":{"tags":["Study"],"summary":"name 으로 조회하기","operationId":"findByName","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/study/v1/list":{"get":{"tags":["Study"],"summary":"모든 스터디 조회 + 페이징","operationId":"list","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/study/v1/id":{"get":{"tags":["Study"],"summary":"id 로 조회하기","operationId":"findByName_1","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/studyrule/v1/studyrules/{studyruleid}":{"delete":{"tags":["StudyRule-삭제"],"summary":"스터디 규칙 삭제","description":"스터디 규칙을 삭제합니다.","operationId":"delete","parameters":[{"name":"studyruleid","in":"path","description":"삭제하고싶은 StudyRuleId 입력","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}},"/api/my-study/v1":{"delete":{"tags":["My Study"],"summary":"가입, 초대 거절","operationId":"delete_1","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RsData"}}}}}}}},"components":{"schemas":{"ModifyStudyRuleRequest":{"required":["about","name","ruleId"],"type":"object","properties":{"name":{"type":"string","description":"변경하고싶은 StudyRule 의 이름(name) 작성","example":"이름수정"},"about":{"type":"string","description":"변경하고싶은 StudyRule 의 소개(about) 작성","example":"소개수정"},"ruleId":{"type":"integer","description":"변경하고싶은 StudyRule 의 규칙(RuleId) 입력","format":"int64","example":2}}},"RsDataStudyRuleDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/StudyRuleDto"},"fail":{"type":"boolean"},"success":{"type":"boolean"}}},"StudyResDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"createDate":{"type":"string","format":"date-time"},"modifyDate":{"type":"string","format":"date-time"},"name":{"type":"string","example":"스터디 이름"},"about":{"type":"string","example":"스터디 소개"},"leader":{"type":"string","example":"스터디 리더"},"capacity":{"type":"integer","format":"int32","example":10},"xp":{"type":"integer","format":"int32","example":10},"bronze":{"type":"integer","format":"int32","example":1},"sliver":{"type":"integer","format":"int32","example":1},"gold":{"type":"integer","format":"int32","example":1},"diamond":{"type":"integer","format":"int32","example":1},"ruby":{"type":"integer","format":"int32","example":1},"platinum":{"type":"integer","format":"int32","example":1},"solvedCount":{"type":"integer","format":"int32","example":6}},"description":"StudyRule 과 양방향 맵핑 되어있는 Study 의 정보"},"StudyRuleDto":{"type":"object","properties":{"id":{"type":"integer","description":"StudyRuleId","format":"int64","example":1},"name":{"type":"string","description":"StudyRule 의 이름(name)","example":"이름"},"about":{"type":"string","description":"StudyRule 의 소개(about)","example":"소개"},"study":{"$ref":"#/components/schemas/StudyResDto"},"ruleId":{"type":"integer","description":"StudyRule 이 의존하는 RuleId","format":"int64","example":1}}},"CreateStudyRuleRequest":{"required":["name"],"type":"object","properties":{"name":{"type":"string","description":"StudyRule의 이름","example":"이름"},"about":{"type":"string","description":"StudyRule 소개","example":"소개"},"studyId":{"type":"integer","description":"StudyRule 이 만들어질 스터디의 아이디(StudyId)","format":"int64","example":1},"ruleId":{"type":"integer","description":"StudyRule 이 반영할 규칙 (RuleId)","format":"int64","example":1}}},"CreateStudyRuleResponse":{"type":"object","properties":{"id":{"type":"integer","description":"StudyRuleId","format":"int64","example":1}}},"RsDataCreateStudyRuleResponse":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/CreateStudyRuleResponse"},"fail":{"type":"boolean"},"success":{"type":"boolean"}}},"AddXpReqDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"xp":{"type":"integer","format":"int32","example":10}}},"RsData":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"type":"object"},"fail":{"type":"boolean"},"success":{"type":"boolean"}}},"UpdateReqDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"name":{"type":"string","example":"스터디 이름"},"about":{"type":"string","example":"스터디 소개"},"capacity":{"type":"integer","format":"int32","example":10}}},"UpdateLeaderReqDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"leader":{"type":"string","example":"스터디 리더"}}},"CreateReqDto":{"type":"object","properties":{"member":{"type":"integer","format":"int64","example":1},"name":{"type":"string","example":"스터디 이름"},"about":{"type":"string","example":"스터디 소개"},"leader":{"type":"string","example":"스터디 리더"},"capacity":{"type":"integer","format":"int32","example":10}}},"ModifyMsgDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1},"msg":{"type":"string","example":"업데이트 메시지"}}},"JoinMyStudyReqDto":{"type":"object","properties":{"study":{"type":"integer","format":"int64","example":1},"member":{"type":"integer","format":"int64","example":1},"msg":{"type":"string","example":"가입 요청 메시지"}}},"InviteMyStudyReqDto":{"type":"object","properties":{"study":{"type":"integer","format":"int64","example":1},"inviter":{"type":"integer","format":"int64","example":1},"invitee":{"type":"integer","format":"int64","example":1},"msg":{"type":"string","example":"초대 메시지"}}},"AcceptDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64","example":1}}},"RsDataListStudyRuleDto":{"type":"object","properties":{"resultCode":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/StudyRuleDto"}},"fail":{"type":"boolean"},"success":{"type":"boolean"}}}}}}